backend/requirements.txt
# Create a new requirements file with compatible versions
cat > requirements-fixed.txt << EOF
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.4
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
transformers==4.30.0
huggingface_hub==0.16.4
torch==2.0.1
PyPDF2==3.0.1
pdfplumber==0.10.3
nltk==3.8.1
spacy==3.6.1
redis==4.6.0
celery==5.3.1
gunicorn==21.2.0
joblib==1.3.2
tqdm==4.65.0
faiss-cpu==1.7.4
EOF

# Install from the fixed requirements
pip install -r requirements-fixed.txt